code/pyproject.toml
[tool.poetry]
name = "backen-product"
version = "0.1.0"
description = "backend-dev"
authors = ["s2rg34-m4r2n4"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
passlib = "^1.7.4"
pydantic = "^2.8.2"
pytest = "^8.3.1"
python-dotenv = "^1.0.1"
uvicorn = "^0.30.3"
jose = "^1.0.0"
python-multipart = "^0.0.9"
uuid = "^1.30"
semgrep = "^1.95.0"
structlog = "^24.4.0"
redis = {extras = ["asyncio"], version = "^5.2.0"}
asyncio = "^3.4.3"
pymongo = {extras = ["srv"], version = "^4.10.1"}
celery = {extras = ["mongodb"], version = "^5.4.0"}
pytest-mock = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
